[FIX] move to separate header file
