+++ src/openms/source/FORMAT/SqliteConnector.cpp
@@ -33,6 +33,7 @@
 // --------------------------------------------------------------------------
 
 #include <OpenMS/FORMAT/SqliteConnector.h>
+#include <OpenMS/FORMAT/SqliteHelper.h>
 
 #include <OpenMS/CONCEPT/Exception.h>
 
@@ -184,166 +185,5 @@ namespace OpenMS
     sqlite3_finalize(stmt);
   }
 
-  namespace Internal
-  {
-    namespace SqliteHelper
-    {
-
-      template <> bool extractValue<double>(double* dst, sqlite3_stmt* stmt, int pos) //explicit specialization
-      {
-        if (sqlite3_column_type(stmt, pos) != SQLITE_NULL)
-        {
-          *dst = sqlite3_column_double(stmt, pos);
-          return true;
-        }
-        return false;
-      }
-
-      template <> bool extractValue<int>(int* dst, sqlite3_stmt* stmt, int pos) //explicit specialization
-      {
-        if (sqlite3_column_type(stmt, pos) != SQLITE_NULL)
-        {
-          *dst = sqlite3_column_int(stmt, pos);
-          return true;
-        }
-        return false;
-      }
-      template <> bool extractValue<Int64>(Int64* dst, sqlite3_stmt* stmt, int pos) //explicit specialization
-      {
-        if (sqlite3_column_type(stmt, pos) != SQLITE_NULL)
-        {
-          *dst = sqlite3_column_int64(stmt, pos);
-          return true;
-        }
-        return false;
-      }
-
-      template <> bool extractValue<String>(String* dst, sqlite3_stmt* stmt, int pos) //explicit specialization
-      {
-        if (sqlite3_column_type(stmt, pos) != SQLITE_NULL)
-        {
-          *dst = String(reinterpret_cast<const char*>(sqlite3_column_text(stmt, pos)));
-          return true;
-        }
-        return false;
-      }
-
-      template <> bool extractValue<std::string>(std::string* dst, sqlite3_stmt* stmt, int pos) //explicit specialization
-      {
-        if (sqlite3_column_type(stmt, pos) != SQLITE_NULL)
-        {
-          *dst = std::string(reinterpret_cast<const char*>(sqlite3_column_text(stmt, pos)));
-          return true;
-        }
-
-        return false;
-      }
-
-      SqlState nextRow(sqlite3_stmt* stmt, SqlState current)
-      {
-        if (current != SqlState::SQL_ROW)
-        { // querying a new row after the last invocation gave 'SQL_DONE' might loop around
-          // to the first entry and give an infinite loop!!!
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Sql operation requested on SQL_DONE/SQL_ERROR state. This should never happen. Please file a bug report!");
-        }
-        int rc = sqlite3_step(stmt);
-        if (rc == SQLITE_ROW)
-        {
-          return SqlState::SQL_ROW;
-        }
-        if (rc == SQLITE_DONE)
-        {
-          return SqlState::SQL_DONE;
-        }
-        if (rc == SQLITE_ERROR)
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Sql operation failed with SQLITE_ERROR!");
-        }
-        if (rc == SQLITE_BUSY)
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Sql operation failed with SQLITE_BUSY!");
-        }
-        if (rc == SQLITE_MISUSE)
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Sql operation failed with SQLITE_MISUSE!");
-        }
-        throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Sql operation failed with unexpected error code!");
-      }
-
-      /// Special case: store integer in a string data value
-      bool extractValueIntStr(String* dst, sqlite3_stmt* stmt, int pos)
-      {
-        if (sqlite3_column_type(stmt, pos) == SQLITE_INTEGER)
-        {
-          *dst = sqlite3_column_int(stmt, pos);
-          return true;
-        }
-        return false;
-      }
-
-      double extractDouble(sqlite3_stmt* stmt, int pos)
-      {
-        double res;
-        if (!extractValue<double>(&res, stmt, pos)) 
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Conversion of column " + String(pos) + " to double failed");
-        }
-        return res;
-      }
-
-      float extractFloat(sqlite3_stmt* stmt, int pos)
-      {
-        double res; // there is no sqlite3_column_float.. so we extract double and convert
-        if (!extractValue<double>(&res, stmt, pos))
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Conversion of column " + String(pos) + " to double/float failed");
-        }
-        return (float)res;
-      }
-
-      int extractInt(sqlite3_stmt* stmt, int pos)
-      {
-        int res;
-        if (!extractValue<int>(&res, stmt, pos))
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Conversion of column " + String(pos) + " to int failed");
-        }
-        return res;
-      }
-
-      Int64 extractInt64(sqlite3_stmt* stmt, int pos)
-      {
-        Int64 res;
-        if (!extractValue<Int64>(&res, stmt, pos))
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Conversion of column " + String(pos) + " to Int64 failed");
-        }
-        return res;
-      }
-
-      String extractString(sqlite3_stmt* stmt, int pos)
-      {
-        String res;
-        if (!extractValue<String>(&res, stmt, pos))
-        {
-          throw Exception::SqlOperationFailed(__FILE__, __LINE__, OPENMS_PRETTY_FUNCTION, "Conversion of column " + String(pos) + " to String failed");
-        }
-        return res;
-      }
-
-      char extractChar(sqlite3_stmt* stmt, int pos)
-      {
-        return extractString(stmt, pos)[0];
-      }
-
-      bool extractBool(sqlite3_stmt* stmt, int pos)
-      {
-        return extractInt(stmt, pos) != 0;
-      }
-
-    }
-  }
-
 }
-
